Use trivially amended branch on blacklight_range_limit
We're using git instead of a rubygems release because the
`blacklight_range_limit` branch is too old to pass CI or even build
(before the trivial renaming of variables).  This application should
upgrade from 2.x to 5.x or 6.x, or it should dedicate resources to
properly maintaning the CI of the 2.x branch.  Right now I'm betting on
the former.

Tl;dr: This fixes 5 of the 7 remaining CI errors.